comment the shit out of it
For some reason it still doesn't work. My only hope is that it only doesn't work when there are angles >90° or <-90° because those are impossible anyway. we need to prevent them from appearing yea but maybe we should focus on that now and forget that it's not working for other angles even though it should...