fix it finally
all the logic for the algorithms and calculations were correct. the only error was that I told FABRIK the shoulder is movable when it is in fact always directly above the base with no possibility to rotate (except in z which doesn't affect the position).

It was a 7 line change (within 250-290).

This took way too long.